+++ tests/test_copy_command.py
@@ -17,7 +17,12 @@ creds = (
 )
 
 
-tbl = sa.Table('t1', sa.MetaData(), schema='schema1')
+tbl = sa.Table(
+    't1', sa.MetaData(),
+    sa.Column('col1', sa.Unicode()),
+    sa.Column('col2', sa.Unicode()),
+    schema='schema1'
+)
 tbl2 = sa.Table('t1', sa.MetaData())
 
 
@@ -34,7 +39,7 @@ def test_basic_copy_case():
     """ % creds
 
     copy = dialect.CopyCommand(
-        table=tbl,
+        tbl,
         data_location='s3://mybucket/data/listing/',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key,
@@ -59,7 +64,7 @@ def test_format():
     TRUNCATECOLUMNS
     """ % creds
     copy = dialect.CopyCommand(
-        table=tbl2,
+        tbl2,
         data_location='s3://mybucket/data/listing/',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key,
@@ -77,7 +82,7 @@ def test_invalid_format():
     t = sa.Table('t1', sa.MetaData(), schema='schema1')
     with pytest.raises(ValueError):
         dialect.CopyCommand(
-            table=t,
+            t,
             data_location='s3://bucket',
             access_key_id=access_key_id,
             secret_access_key=secret_access_key,
@@ -96,7 +101,7 @@ def test_compression():
     TRUNCATECOLUMNS
     """ % creds
     copy = dialect.CopyCommand(
-        table=tbl,
+        tbl,
         data_location='s3://mybucket/data/listing/',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key,
@@ -113,7 +118,7 @@ def test_compression():
 def test_invalid_compression():
     with pytest.raises(ValueError):
         dialect.CopyCommand(
-            table=tbl,
+            tbl,
             data_location='s3://bucket/of/joy',
             access_key_id=access_key_id,
             secret_access_key=secret_access_key,
@@ -135,7 +140,7 @@ def test_ascii_nul_as_redshift_null():
     TRUNCATECOLUMNS
     """ % creds
     copy = dialect.CopyCommand(
-        table=tbl,
+        tbl,
         data_location='s3://mybucket/data/listing/',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key,
@@ -150,23 +155,40 @@ def test_ascii_nul_as_redshift_null():
     assert clean(expected_result) == clean(compile_query(copy))
 
 
-def test_json_upload():
+def test_json_upload_with_manifest_ordered_columns():
     expected_result = """
-    COPY schema1.t1 FROM 's3://mybucket/data/listing/'
+    COPY schema1.t1 (col1, col2) FROM 's3://mybucket/data/listing.manifest'
     WITH CREDENTIALS AS '%s'
-    FORMAT AS JSON AS 'auto'
+    FORMAT AS JSON AS 's3://mybucket/data/jsonpath.json'
     GZIP
+    MANIFEST
     ACCEPTANYDATE
     TIMEFORMAT AS 'auto'
     """ % creds
     copy = dialect.CopyCommand(
-        table=tbl,
-        data_location='s3://mybucket/data/listing/',
+        [tbl.c.col1, tbl.c.col2],
+        data_location='s3://mybucket/data/listing.manifest',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key,
+        manifest=True,
         format='JSON',
+        path_file='s3://mybucket/data/jsonpath.json',
         compression='GZIP',
         time_format='auto',
         accept_any_date=True,
     )
     assert clean(expected_result) == clean(compile_query(copy))
+
+
+def test_different_tables():
+    metdata = sa.MetaData()
+    t1 = sa.Table('t1', metdata, sa.Column('col1', sa.Unicode()))
+    t2 = sa.Table('t2', metdata, sa.Column('col1', sa.Unicode()))
+    with pytest.raises(ValueError):
+        dialect.CopyCommand(
+            [t1.c.col1, t2.c.col1],
+            data_location='s3://bucket',
+            access_key_id=access_key_id,
+            secret_access_key=secret_access_key,
+            format='CSV'
+        )